requirements.txt
streamlit
pandas
requests
altair
flask
pymongo
flask-cors
streamlit_autorefresh
google-genai
scikit-learn
twilio